app-004_bot-discord/requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py==1.5.1
idna==2.10
idna-ssl==1.1.0
multidict==4.7.6
python-dotenv==0.15.0
typing-extensions==3.7.4.3
yarl==1.5.1
